refactor: move init operation out of start gateway command to fit sidecar mode in the future
