+++ .gitignore
@@ -12,3 +12,4 @@ pkg/registry/zookeeper-4unittest/contrib/fatjar
 
 coverage.txt
 
+/vendor/
