--- cmd/pixiu/control.go
@@ -1,152 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package main
-
-import (
-	"os"
-	"runtime"
-	"strconv"
-	"time"
-)
-
-import (
-	"github.com/spf13/cobra"
-)
-
-import (
-	"github.com/apache/dubbo-go-pixiu/pkg/common/constant"
-	"github.com/apache/dubbo-go-pixiu/pkg/config"
-	"github.com/apache/dubbo-go-pixiu/pkg/logger"
-	"github.com/apache/dubbo-go-pixiu/pkg/pixiu"
-)
-
-var (
-	flagToLogLevel = map[string]string{
-		"trace":    "TRACE",
-		"debug":    "DEBUG",
-		"info":     "INFO",
-		"warning":  "WARN",
-		"error":    "ERROR",
-		"critical": "FATAL",
-	}
-
-	configPath    string
-	apiConfigPath string
-	logConfigPath string
-	logLevel      string
-
-	// CURRENTLY USELESS
-	logFormat string
-
-	limitCpus string
-
-	// Currently default set to false, wait for up coming support
-	initFromRemote = false
-
-	startCmd = &cobra.Command{
-		Use:   "dubbogo pixiu",
-		Short: "Dubbogo pixiu is a lightweight gateway.",
-		Long: "dubbo-go-pixiu is a gateway that mainly focuses on providing gateway solution to your Dubbo and RESTful \n" +
-			"services. It supports HTTP-to-Dubbo and HTTP-to-HTTP proxy and more protocols will be supported in the near \n" +
-			"future." +
-			"(c) " + strconv.Itoa(time.Now().Year()) + " Dubbogo",
-		Version: Version,
-		PreRun: func(cmd *cobra.Command, args []string) {
-			initDefaultValue()
-		},
-		Run: func(cmd *cobra.Command, args []string) {
-			err := logger.InitLog(logConfigPath)
-			if err != nil {
-				// cause `logger.InitLog` already handle init failed, so just use logger to log
-				logger.Warnf("[startCmd] failed to init log config file")
-			}
-
-			if level, ok := flagToLogLevel[logLevel]; ok {
-				logger.SetLoggerLevel(level)
-			} else {
-				logger.Warnf("[startCmd] logLevel is invalid, set log level to default: %s", constant.DefaultLogLevel)
-				logger.SetLoggerLevel(flagToLogLevel[constant.DefaultLogLevel])
-			}
-
-			bootstrap := config.Load(configPath)
-			// get api config from meta data center
-			if bootstrap.GetAPIMetaConfig() != nil {
-				if _, err := config.LoadAPIConfig(bootstrap.GetAPIMetaConfig()); err != nil {
-					logger.Warnf("[startCmd] failed to get api meta config, %s", err.Error())
-				}
-			}
-
-			// get api config from local file
-			if !initFromRemote {
-				if _, err := config.LoadAPIConfigFromFile(apiConfigPath); err != nil {
-					logger.Errorf("[startCmd] failed to load api config from file, %s", err.Error())
-				}
-			}
-
-			limitCpuNumber, err := strconv.ParseInt(limitCpus, 10, 64)
-			if err != nil {
-				logger.Errorf("[startCmd] failed to get limit cpu number, %s", err.Error())
-			}
-			if limitCpuNumber <= 0 {
-				runtime.GOMAXPROCS(runtime.NumCPU())
-			} else {
-				runtime.GOMAXPROCS(int(limitCpuNumber))
-			}
-
-			pixiu.Start(bootstrap)
-		},
-	}
-)
-
-// init Init startCmd
-func init() {
-	startCmd.PersistentFlags().StringVarP(&configPath, constant.ConfigPathKey, "c", os.Getenv(constant.EnvDubbogoPixiuConfig), "Load configuration from `FILE`")
-	startCmd.PersistentFlags().StringVarP(&apiConfigPath, constant.ApiConfigPathKey, "a", os.Getenv(constant.EnvDubbogoPixiuApiConfig), "Load api configuration from `FILE`")
-	startCmd.PersistentFlags().StringVarP(&logConfigPath, constant.LogConfigPathKey, "g", os.Getenv(constant.EnvDubbogoPixiuLogConfig), "Load log configuration from `FILE`")
-	startCmd.PersistentFlags().StringVarP(&logLevel, constant.LogLevelKey, "l", os.Getenv(constant.EnvDubbogoPixiuLogLevel), "dubbogo pixiu log level, trace|debug|info|warning|error|critical")
-	startCmd.PersistentFlags().StringVarP(&limitCpus, constant.LimitCpusKey, "m", os.Getenv(constant.EnvDubbogoPixiuLimitCpus), "dubbogo pixiu schedule threads count")
-	startCmd.PersistentFlags().StringVarP(&logFormat, constant.LogFormatKey, "f", os.Getenv(constant.EnvDubbogoPixiuLogFormat), "dubbogo pixiu log format, currently useless")
-
-}
-
-// initDefaultValue If not set both in args and env, set default values
-func initDefaultValue() {
-	if configPath == "" {
-		configPath = constant.DefaultConfigPath
-	}
-
-	if apiConfigPath == "" {
-		apiConfigPath = constant.DefaultApiConfigPath
-	}
-
-	if logConfigPath != "" {
-		logConfigPath = constant.DefaultLogConfigPath
-	}
-
-	if logLevel == "" {
-		logLevel = constant.DefaultLogLevel
-	}
-
-	if limitCpus == "" {
-		limitCpus = constant.DefaultLimitCpus
-	}
-
-	if logFormat != "" {
-		logFormat = constant.DefaultLogFormat
-	}
-}
