+++ start.sh
@@ -26,4 +26,4 @@ PROJECT_HOME=${PROJECT_HOME}"/"
 export CONF_CONSUMER_FILE_PATH=${PROJECT_HOME}"conf/client.yml"
 export APP_LOG_CONF_FILE=${PROJECT_HOME}"conf/log.yml"
 
-./dubbo-go-pixiu ${args}
+./dubbo-go-pixiu start-gateway ${args}
